requirements.txt
anthropic>=0.40.0
beautifulsoup4>=4.12.0
feedparser>=6.0.11
google-api-python-client>=2.150.0
google-auth>=2.35.0
google-auth-oauthlib>=1.2.0
httpx>=0.27.0
pydantic>=2.9.0
python-dateutil>=2.9.0
python-dotenv>=1.0.0
pyyaml>=6.0.2
replicate>=0.34.0
requests>=2.32.0
tenacity>=9.0.0
